move opt to its own mod
