+++ src/opt.rs
@@ -1,0 +1,82 @@
+use std::env;
+
+use structopt::StructOpt;
+
+#[derive(StructOpt)]
+#[structopt(name = "taskn", about = "Taskwarrior task annotation helper")]
+struct ProtoOpt {
+    /// The editor used to open task notes. If unset, taskn will attempt to use $EDITOR. If $EDITOR
+    /// is also unset, taskn will use vi.
+    #[structopt(long)]
+    editor: Option<String>,
+
+    /// The file format used for task notes.
+    #[structopt(long, default_value = "md")]
+    file_format: String,
+
+    /// The directory in which task notes are placed. If the directory does not already exist,
+    /// taskn will create it.
+    #[structopt(long, default_value = "~/.taskn")]
+    root_dir: String,
+
+    #[structopt(long, default_value = "edit")]
+    command: String,
+
+    args: Vec<String>,
+}
+
+impl ProtoOpt {
+    fn into_opt(self) -> Opt {
+        let editor = if let Some(editor) = self.editor {
+            editor
+        } else if let Ok(editor) = env::var("EDITOR") {
+            editor
+        } else {
+            "vi".to_string()
+        };
+
+        let root_dir = shellexpand::tilde(&self.root_dir).to_string();
+
+        Opt {
+            editor,
+            file_format: self.file_format,
+            root_dir,
+            command: self.command,
+            args: self.args,
+        }
+    }
+}
+
+pub struct Opt {
+    pub editor: String,
+    pub file_format: String,
+    pub root_dir: String,
+    pub command: String,
+    pub args: Vec<String>,
+}
+
+impl Opt {
+    fn from_proto_opt(proto_opt: ProtoOpt) -> Self {
+        let editor = if let Some(editor) = proto_opt.editor {
+            editor
+        } else if let Ok(editor) = env::var("EDITOR") {
+            editor
+        } else {
+            "vi".to_string()
+        };
+
+        let root_dir = shellexpand::tilde(&proto_opt.root_dir).to_string();
+
+        Opt {
+            editor,
+            file_format: proto_opt.file_format,
+            root_dir,
+            command: proto_opt.command,
+            args: proto_opt.args,
+        }
+    }
+
+    pub fn from_args() -> Self {
+        Self::from_proto_opt(ProtoOpt::from_args())
+    }
+}
